refactor: The Repository was edited so that it is functional

+++ src/main/java/com/dominickchavarria/kinlapp/repository/ProductoRepository.java
@@ -5,7 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface ProductoRepository extends JpaRepository<Producto, Integer> {
-    List<Producto> findByEstado(int estado);
-
+public interface ProductoRepository extends JpaRepository<Producto, Long>{
+    List<Producto> findByEstado(Long estado);
 }
